% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_county_income}
\alias{census_county_income}
\title{2018 population, density, and median income estimates from U.S. Census}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3220 rows and 6 columns.
}
\usage{
census_county_income
}
\description{
2018 population, density, and median income estimates from U.S. Census
}
\keyword{datasets}
