% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_state_pop}
\alias{census_state_pop}
\title{Google Mobility Reports
2018 population, density, and median income estimates from U.S. Census}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 4 columns.
}
\usage{
census_state_pop
}
\description{
Google Mobility Reports
2018 population, density, and median income estimates from U.S. Census
}
\keyword{datasets}
