% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidtracking}
\alias{covidtracking}
\title{COVID-19 data from COVID Tracking Project, current as of 2020-04-29}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3041 rows and 26 columns.
}
\source{
\url{covidtracking.com}
}
\usage{
covidtracking
}
\description{
COVID-19 data from COVID Tracking Project, current as of 2020-04-29
}
\keyword{datasets}
