% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csse_us}
\alias{csse_us}
\title{Johns Hopkins CSSE COVID-19 U.S. county-level data, current as of 2020-04-27}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 313152 rows and 14 columns.
}
\source{
\url{github.com/CSSEGISandData/COVID-19}
}
\usage{
csse_us
}
\description{
Johns Hopkins CSSE COVID-19 U.S. county-level data, current as of 2020-04-27
}
\keyword{datasets}
