% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat5730covid19-package.R
\docType{package}
\name{stat5730covid19-package}
\alias{stat5730covid19}
\alias{stat5730covid19-package}
\title{stat5730covid19: COVID-19 Data and Tools for STAT 5730}
\description{
Tools to import COVID-19 data from JHU CSSE, Covid Tracking Project, NY Times, and Google for the students of STAT 5730.
}
\author{
\strong{Maintainer}: Vincent Vu \email{vqv@stat.osu.edu} (\href{https://orcid.org/0000-0002-4689-0497}{ORCID})

}
\keyword{internal}
