% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_orders}
\alias{state_orders}
\title{U.S. state executive orders data from STAT 5730, current as of 2020-04-27}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 4 columns.
}
\usage{
state_orders
}
\description{
U.S. state executive orders data from STAT 5730, current as of 2020-04-27
}
\keyword{datasets}
