% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csse_global}
\alias{csse_global}
\title{Johns Hopkins CSSE COVID-19 international data, current as of 2020-05-01}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27500 rows and 8 columns.
}
\source{
\url{github.com/CSSEGISandData/COVID-19}
}
\usage{
csse_global
}
\description{
Johns Hopkins CSSE COVID-19 international data, current as of 2020-05-01
}
\keyword{datasets}
